Kernel: Don't send SIGCHLD for not fully initialized processes

+++ Kernel/Tasks/Process.cpp
@@ -908,8 +908,9 @@ void Process::finalize()
     m_state.store(State::Dead, AK::MemoryOrder::memory_order_release);
 
     {
-        if (auto parent_process = Process::from_pid_ignoring_process_lists(ppid())) {
-            if (parent_process->is_user_process() && (parent_process->m_signal_action_data[SIGCHLD].flags & SA_NOCLDWAIT) != SA_NOCLDWAIT)
+        if (is_fully_initialized()) {
+            auto parent_process = Process::from_pid_ignoring_process_lists(ppid());
+            if (parent_process && parent_process->is_user_process() && (parent_process->m_signal_action_data[SIGCHLD].flags & SA_NOCLDWAIT) != SA_NOCLDWAIT)
                 (void)parent_process->send_signal(SIGCHLD, this);
         }
     }
